refactor(results): moved the getResults function with the API call to lib/istexApi.ts
